PPF-366 - Progress on approve integration E2E test
