+++ e2e/tests/integrations/integrator/request-activation.test.ts
@@ -1,5 +1,6 @@
 import { expect, test } from "@playwright/test";
 import { fakerNL_BE as faker } from "@faker-js/faker";
+import { requestActivationAsIntegrator } from "./request-activation";
 
 test("As an integrator I can request activation via organization details", async ({
   page,
@@ -50,29 +51,5 @@ test("As an integrator I can request activation via organization details", async
   const newIntegrationId = url.split("/").pop();
 
   // Request activation
-  await page.goto(`/nl/integraties/${newIntegrationId}`);
-  await page
-    .getByRole("button", { name: "Activatie aanvragen" })
-    .nth(1)
-    .click();
-  await page
-    .getByLabel("Naam bedrijf, organisatie of")
-    .nth(1)
-    .fill(faker.company.name());
-  await page
-    .getByLabel("Straat en nummer")
-    .nth(1)
-    .fill(faker.location.streetAddress());
-  await page.getByLabel("Postcode").nth(1).fill(faker.location.zipCode("####"));
-  await page.getByLabel("Gemeente").nth(3).fill(faker.location.city());
-  await page
-    .getByLabel("BTW of ondernemingsnummer")
-    .nth(1)
-    .fill("BE 0475 250 609");
-  await page
-    .getByLabel("E-mail boekhouding")
-    .nth(1)
-    .fill(faker.internet.email());
-  await page.getByRole("button", { name: "Bevestigen" }).nth(1).click();
-  await expect(page.getByText("Actief").nth(1)).toBeVisible();
+  await requestActivationAsIntegrator(page, newIntegrationId!);
 });
